Game Of Life Implementation v0.8
Implementation of a game of life board generator according to John Conway. 
This first version is intended to generate a board with a random or inputed pattern and to calculate it's generations according to the rules. The output is printed and saved to a png image with pretty colors.
Implemented rules : https://en.wikipedia.org/wiki/Conway%27s_Game_of_Life